split operator functions into their own files
