+++ packages/core/src/operator/operator-runtime/index.ts
@@ -1,5 +1,13 @@
 export * from "./operatorState.js";
-export * from "./processNewChallenge.js";
 export * from "./updateNodeLicenseStatus.js";
-export * from "./createAssertionHashAndCheckPayout.js";
-export * from "./calculateBoostFactor.js";
+export * from "./operator-v1/bootOperatorRuntime.js";
+export * from "./findSubmissionOnSentryKey.js";
+export * from "./operator-v1/processNewChallenge.js";
+export * from "./operator-v1/processClosedChallenges.js";
+export * from "./operator-v1/processPastChallenges.js";
+export * from "./operator-v1/createAssertionHashAndCheckPayout.js";
+export * from "./operator-v1/calculateBoostFactor.js";
+export * from "./operator-v1/processClosedChallenges.js";
+export * from "./operator-v1/loadOperatorKeysFromGraph.js";
+export * from "./operator-v1/loadOperatorKeysFromRPC.js";
+
