+++ packages/core/src/operator/operator-runtime/operator-v1/createAssertionHashAndCheckPayout.ts
